statistics

+++ ServerImpl/Server/IServer.cs
@@ -133,6 +133,6 @@ namespace Server
 
         Tuple<string, List<Test>> getGroupTests(int adminId, string group);
 
-        Tuple<string, List<Tuple<string, int>>> getGrades(int adminId, int testId, string groupName);
+        Tuple<string, List<Tuple<string, double>>> getGrades(int adminId, int testId, string groupName);
     }
 }
