+++ ServerImpl/communication/Models/TestStatistics/TestStatisticsData.cs
@@ -7,7 +7,7 @@ namespace communication.Models.TestStatistics
 {
     public class TestStatisticsData
     {
-        public List<Tuple<string, int>> gradesInTest { get; set; }
+        public List<Tuple<string, double>> gradesInTest { get; set; }
 
     }
 }
